Adding comments and minor fixes
